- Implementa Log; - Implementa o Search; - Implementa Dtos;

+++ MyMovies/Controllers/BaseController.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using IdentityModel;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MyMovies.Api.Filters;
 using MyMovies.Entities;
+using MyMovies.Entities.Dto;
 using MyMovies.Repositories.Database.Interfaces;
 using MyMovies.Repositories.Interfaces;
 
@@ -10,57 +12,93 @@ namespace MyMovies.Api.Controllers
     [ApiController]
     [Route("[controller]")]
     [Authorize]
-    public abstract class BaseController<T> : Controller
+    public abstract class BaseController<T, TDto> : Controller
         where T : Movie
+        where TDto : IDto
     {
-        private readonly IDatabaseRepository<T> repository;
+        private readonly IDatabaseRepository<T, TDto> repository;
 
-        public BaseController(IDatabaseRepository<T> repository)
+        public BaseController(IDatabaseRepository<T, TDto> repository)
         {
             this.repository = repository;
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public T Get(long id)
         {
             return repository.Read(id);
         }
 
-        [HttpGet]
+        [HttpPost]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [Route("query")]
-        public IEnumerable<T> List([FromBody] T entity)
+        public async Task<IActionResult> List([FromBody] TDto entity)
         {
-            return repository.Search(entity);
+            var ret = repository.Search(entity);
+
+            if(ret == null || ret.Count == 0)
+            {
+                return NotFound();
+            }
+            return Ok(ret);
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [Route("listAll")]
-        public IEnumerable<T> ListAll()
+        public async Task<IActionResult> ListAll()
         {
-            return repository.ReadAll();
+            var ret = repository.ReadAll();
+
+            if (ret == null || ret.Count == 0)
+            {
+                return NotFound();
+            }
+            return Ok(ret);
         }
 
         [HttpPut]
-        [CustomActionFilterEndpoint("Put")]
-        public T Put(long id, T entity)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [CustomActionFilterEndpoint]
+        public async Task<IActionResult> Put(long id, T entity)
         {
             //seta a propriedade Id por reflection pois vem vazia e é read-only
             typeof(T).BaseType.GetProperty("Id")?.SetValue(entity, id, null);
-            return repository.Update(entity);
+            var obj = repository.Update(entity);
+            if (obj == null)
+            {
+                return NotFound();
+            }
+            return Ok(obj);
         }
 
         [HttpPost]
-        [CustomActionFilterEndpoint("Post")]
-        public T Post(T entity)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [CustomActionFilterEndpoint]
+        public async Task<IActionResult> Post(T entity)
         {
-            return repository.Create(entity);
+            return Ok(repository.Create(entity));
         }
 
         [HttpDelete]
-        [CustomActionFilterEndpoint("Delete")]
-        public void Delete(long id)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [CustomActionFilterEndpoint]
+        public async Task<IActionResult> Delete(long id)
         {
+            var obj = repository.Read(id);
+
+            if (obj == null)
+            {
+                return NotFound();
+            }
             repository.Delete(id);
+            return Ok();
         }
     }
 }
